chore: update changed buttons
